Fix main page images
This is a workaroung to avoid missing images on the main page.
The main page is the first page of the book but duplicated in a
index.html at the book root. Unfortunatly, the images contained in this
page are not duplicated. That is why we need to copy them by hand.